Correct type of cit_collision_callback() to match EDMS interface

+++ src/GameSrc/physics.c
@@ -143,7 +143,7 @@ extern errtype (*state_generators[])(ObjID id, int x, int y, ObjLocState *ret);
 long old_ticks;
 
 // Collision callback testing....
-void cit_collision_callback(physics_handle C, physics_handle V, int bad, long DATA1, long DATA2, fix location[3]);
+void cit_collision_callback(physics_handle C, physics_handle V, int32_t bad, int32_t DATA1, int32_t DATA2, fix location[3]);
 void cit_awol_callback(physics_handle caller);
 void cit_sleeper_callback(physics_handle caller);
 void cit_autodestruct_callback(physics_handle caller);
@@ -1564,7 +1564,7 @@ void terrain_object_collide(physics_handle src, ObjID target) {
     collide_objects(target, hit_obj, 0);
 }
 
-void cit_collision_callback(physics_handle C, physics_handle V, int bad, long DATA1, long DATA2, fix location[3]) {
+void cit_collision_callback(physics_handle C, physics_handle V, int32_t bad, int32_t DATA1, int32_t DATA2, fix location[3]) {
     ObjID collision;
     ObjID victim;
